work with cart and admin orders

+++ app/models/option.rb
@@ -1,8 +1,8 @@
 class Option
   include Mongoid::Document
-  field :mass, type: Integer
-  field :volume, type: Integer
-  field :price, type: Integer
+  field :mass, type: Float
+  field :volume, type: Float
+  field :price, type: Float
   field :amount, type: Integer
   field :packing, type: String
   belongs_to :product
